+++ changelog.md
@@ -1,3 +1,8 @@
+## 21.0.0
+- removed twitter bootstrap
+- autocomplete - disabled & readOnly style fix
+- select - disabled & readOnly style fix
+
 ## 20.2.0
 - backdrop - style fixes
 - dateselect - use new grid